+++ packages/core/src/Engine.ts
@@ -5,8 +5,8 @@ import { ShaderData } from './shader/ShaderData';
 import { ShaderMacroCollection } from './shader/ShaderMacroCollection';
 import { ShaderMacro } from './shader/ShaderMacro';
 import { ShaderProperty } from './shader/ShaderProperty';
-import { Event } from './base/Event';
-import { EventDispatcher } from './base/EventDispatcher';
+import { Event } from './base/event';
+import { EventDispatcher } from './base/eventDispatcher';
 import { Time } from './base/Time';
 import { RenderContext } from './renderer/RenderContext';
 import { RenderPipeline } from './renderer/RenderPipeline';
