+++ packages/core/src/index.ts
@@ -31,13 +31,13 @@ export * from './utils';
 export * from './interface';
 
 // 基础组件
-export * from './base/Event';
-export * from './base/EventDispatcher';
+export * from './base/event';
+export * from './base/eventDispatcher';
 export * from './base/Canvas';
 export * from './base/Time';
 export * from './base/ObjectPool';
 export * from './base/ObjectPoolManager';
-export * from './base/EnhancedEventDispatcher';
+export * from './base/eventDispatcher';
 
 // 引擎核心
 export * from './Engine';
